refactor: query 분리
