+++ src/types/tab.ts
@@ -1,0 +1,7 @@
+interface CurrentTab {
+  title: string;
+  currentUrl: string;
+  favicon: string;
+}
+
+export default CurrentTab;
